implement v_from_i wrapper
* add new test_i_from_v_from_i which recycles the test fixture from
v_from_i which has more reasonable test values that are not way down in
quadrant 4, it uses the old i_from_v(method='lambertw') to recalculate
I so it has the write shape, size and dtype
* set method='lambertw' for old i_from_v test, because since the Voc is
like 8V, and the test voltage is 40V, the expected current is -300A,
which is unlikely, and I doubt very very very much that this calculated
value is anywhere near the actual current. Ha, ha, ha, 300A, that is so
funny!
* change PVSystem_i_from_v test to use the v_from_i fixture values which
actually make some sense
* set v_from_i inside singlediode, to use the method='lambertw' when
given
* add default method='' to v_from_i, add docstring
* add docstring to i_from_v
* add section to size and shape
* wrap i_from_v_fun with returns_nan(), defaults to ValueError and
RuntimeError
* add returns_nan() decorator to way_faster