Add functionality for reading EPW weather files (#677)
* Add functionality for reading EPW weather files

Adds EPW reader to iotools.tmy and includes files for testing

* Include documentation for read_epw

* read_epw is now in a separate module instead of part of tmy.py

includes a couple of other small fixes and updated testing files

* Update epw.py

Remove line that was only meant for debugging

* Force date format

This commit forces the data format and now pads a zero in front of the 'hour' column. I believe that this solves the Python 2.7 issue, but Travis testing needs to confirm this

* Remove debugging statement

* Testing dir update and new code structure for localizing index

note .dt option in localizing. see: https://stackoverflow.com/questions/26089670/unable-to-apply-methods-on-timestamps-using-series-built-ins

* Update index creation following Will's suggestion and cleaned up code

Updates code for index creation in epw.py
Improves docstring for clarity
Cleans trailing spaces, missing whitespace, etc. identified by pycodestyle in epw.py and test_epw.py

* Changed indexing of last hour of the year

Last hour of the year led to an error in the forced year check. It was because of a remnant of my quick-and-dirty approach which is no longer necessary using Will's solution.

* Minor change to default input for filename and update of What's New