Remove unnecessary checks in merge().
Returns merge method to its original state with two nested for loops, but without
the checks for the number of dictionaries in the reports argument, or the check
that each element in reports is a dictionary. Reports are only created in the build
method so a test was made to ensure that build always returns a dictionary instead.