refactor bifacial merge, improve merge tests (#747)
* Simplify merge method.

Simplify merge method by using list and dictionary comprehensions instead of
nested for loops. This also avoids the need to check if some of the elements
of the reports arguments are None, and the need to check if reports contains
more than one report. This should also be a small performance improvement.
This change also adds a bit more detail to the docstring for the merge method.
The creation of a new empty report is also modified to no longer use an
intermediary list of keys.

* Update what's new doc.

* Update whatsnew file syntax.

* Rename whatsnew file to change format to rst.

* Fix linting errors.

* Remove pandas and numpy version reqs

* Update v0.7.1.rst

* Revert "Update v0.7.1.rst"

This reverts commit d131657b18b4c309f709fc3b3931b1d7c51471f3.

* Revert "Remove pandas and numpy version reqs"

This reverts commit 84d718949289c82d6d3e1eb7b0b0fa05c413adf0.

* Reword elements in merge() comprehension.

* Fix handling of None values in merge.

If one of the values in the reports argument that merge() takes is a None,
this is now properly handled by being dropped.

* Remove unnecessary checks in merge().

Returns merge method to its original state with two nested for loops, but without
the checks for the number of dictionaries in the reports argument, or the check
that each element in reports is a dictionary. Reports are only created in the build
method so a test was made to ensure that build always returns a dictionary instead.

* Add tests for merge() and build().

* Fix lint errors.

* Correct import mistake.

Previous commits I made did not import the PVFactorsReportBuilder
class correctly.

* Reinsert type check in merge loop.

* Revert "Reinsert type check in merge loop."

This reverts commit eeb80530ca9b6b1f92a401a44ad80fb43c9e6f7c.

* Fix test_build_1 test.

* Update what's new file.

Add bullet point about added tests for bifacial.py methods.

* Fix merge conflicts in what's new file.

* Revert "Fix merge conflicts in what's new file."

This reverts commit 74c8ebaa5ff81a8202f0a8a4b16571e27d64a6bb.

* Revert "Update what's new file."

This reverts commit 4447e6c5d176a7d10a2a49b64f74b102827b93ce.

* Revert "Rename whatsnew file to change format to rst."

This reverts commit 93d05595b3d88ffa90df89695eb72521dbe536e7.

* Revert "Update whatsnew file syntax."

This reverts commit c969994ff7347779ec27d19318bdbb0374bf7cb3.

* Revert "Update what's new doc."

This reverts commit 55d8f08f8a10c94e363e4e9da558a514417459ce.

* Merge what's new updates from upstream.

* Update what's new file.