Refactor PVSystem to use pvsystem.Array internally
Removes array-related fields and adds _array field. All internal references
to array-related attributes now refer to the the corresponding attributes of
the Array class.

It is likely some of the Array attributes will need to be exposed on the
PVSystem instances via @properties; however, any of these instances are not
covered by tests so it will be necessary to carefully work through the other
places where PVsystem is used (ModelChain and child classes) to see what
needs to be done.