Require tuple input in list_or_scalar decorator
After change to returning tuple as lists we need to update the
decortor to check for the correct type. It is tempting to test
for any Iterable here; however, since some of the parameters are
tuples of Series or DataFrame (which are themselves iterabel),
we need to be careful. I believe the best option is to be very
restrictive and expect input to be a tuple.