Replace validation decorator with a private method
The decorator adds too much complexity and room for error in the future.
A validation method that is called once on each parameter that needs to
be the same shape as the `self._arrays` tuple is much easier to read,
understand, and maintain. It also has the benefit of increased flexibility
since it allow validation to be triggered at any point, rather than before
the method is called.