Test ModelChain with multi-array system and no losses
Add a test that verifies the dc side of the ModelChain returns the correct
data for each Array with `aoi_model='no_loss'` and
`spectral_model='no_loss'`.

Test fails because `ModelChain.results.spectral_modifier` and
`ModelChain.results.aoi_modifier` are assigned a scalar value instead
of a tuple. It is very likely we will also see a failure at the
`ModelChain.losses_model()` call in `_run_from_effective_irradiance()`
for the same reason.

At this point we should (maybe) expect a failure at the call to
`ModelChain.ac_model()` since we have not implemented an inverter model
for multiple arrays yet.