Initialize losses parameters corresponding to system._arrays
In the no_losses case, loss parameters are initialized to 1. This
commit initializes a tuple with the same number of elements as the
number of arrays in the system. In cases where losses are being
modeled, these fields are initialized to the result of a PVSystem
method call, meaning they have the correct type and shape. For cases
with no losses we need to take care to manually initialize them
correctly.

After this commit test_run_model_from_irradiance_arrays_no_loss()
fails at the call to PVSystem.pvwatts_ac() which is expected since
we have not implemented inverter models with multiple DC inputs yet.