ModelChain._prepare_temperature accepts multiple DataFrames
If multiple dataframes are passed in a tuple then _prepare_temperature
will look for 'cell_temperature' or 'module_temperature' in each data
frame. For data frames that do not have one of these columns the
ModelChain.temperature_model() is applied to calculate the cell
temperature from the weather data that has already been assigned in
other ModelChain fields.