Don't allow SingleAxisTracker Array with non-None tilt/azimuth
If a list of one Array is given a value error is raised when it has
non-None tilt or azimuth. Rather than just assigning None we raise
an error so that the Array object itself is not modified (since it
could be in use by the user somewhere else). When only parameters
are passed (as opposed to an Array object) we override the values.
Since the Array is created internally and we know it is not in use
elsewhere this isn't a problem.