Separate ac model inference for multiple arrays
Adds an internal _infer_ac_model_multi() method to handle inference
when system.num_arrays > 1. This lets us raise a more usefule error
message in this case.

In addition several _xxxx_params() predicates are added to prevent
duplication of the specific tests on the inverter_params set in
_infer_ac_model_multi(). Seems like over-kill for now, but an
inverter.pvwatts_multi() model is in the works, so this will save
some work when integrating that.