Remove ModelChain._check_consistent_params
Rework test to maintain coverage on inconsistent array parameters,
but expect the error to come from a different place. Mainly this means
that the error message has changed and more input is needed to the
constructor to make sure that execution goes all the way to the
assignment of the temperature model.