Use unwrap=True to simplify ModelChain._singldiode()
This lets us expect a consistent type from the `PVSystem` methods
regardless of the number of arrays. Because of that we can remove
the control-flow that made the function hard to read. This comes
at the cost of needing to unwrap the single value ourselves
before returning, but a single if statement at the end of the
method is much easier to reason about.

Because of the way mocks interact with wrapped functions we need to
move the mock from the object layer to the function layer, or it will
fail because the signature of the wrapped PVSystem methods does not
include the `unwrap` parameter.