Add modelchain.ModelChainResult to Classes section of api.rst
Requires slight modification of sphinx/source/conf.py to address
getting line numbers for the ModelChainResult.__init__ method
which is not explicitly defined (and therefore does not have a line
number).

Made the T TypeVar private so it does not clutter the attribute list
for ModelChainResult.

Add docstring for the PerArray type.