Merged PR 13463: Horizon coefficient should be allowed to go negative in the Perez transposition model
Horizon coefficient should be allowed to go negative in the Perez transposition model

Link to the original paper: https://www.osti.gov/servlets/purl/7024029
Section III.2 states this explicitly for the horizon component:
"(2) The horizon brightening coefficient, F2, is negative for overcast and low E occurrences -- indicative of brightening of the zenithal region of the sky for these conditions. This becomes positive past intermediate conditions and increases substantially with clearness."

Note that this change also impacts results from the DIRINT model (used to generate DNI from GHI).

This change corresponds to revision 8 of the SAM vs PVLib comparison, found here:
https://cleanpower1.sharepoint.com/sites/SolarAnywhere/Shared%20Documents/Power%20Model%20Discussions/PVLib_Comparison_Rev8_FixPvLibPerezModelF2Bound.zip