pvlib.iotools.read_psm3 updates for SAM
The function description states that it reads SAM files, but the SAM files does not need a 'Local Time Zone' in the metadata, as well as the Minute column is optional.

When Minutes are included in the SAM file, it considers the sun position needs to be calculated at that specific point, instead of calculating it for left labeled data (timestamp 10 AM is data left averaged from 10 AM to 11 AM, and SAM places Sun position at 10:30 AM).