Restructure API Reference content to improve left sidebar (#1598)
* split pv_modeling.rst into a directory

The left sidebar in pydata-sphinx-theme reflects document hierarchy, without considering in-page headers.  To get those headers into the sidebar, we have to rework the document hierarchy such that the former in-page headers are now reflected in the document structure.  Not ideal, but it's the only way I've found so far.

* split up effects_on_output and irradiance; use top-level toctree folder

* add dropped entries from merge

* put back ADR entries

* put back astm_e1036